perf(precompile): vendor blake2b_simd for BLAKE2 compression
Vendor the portable and AVX2 BLAKE2b compression functions from
blake2b_simd, replacing the hand-rolled implementations. The vendored
code adds dynamic CPU feature detection for AVX2 (including compile-time
detection via target_feature) and uses the same optimized SIMD
implementations from blake2b_simd, adapted for EIP-152's variable round
count.

This also removes the arrayref dependency usage in the old avx2 module
(the vendored avx2.rs still uses arrayref for the same patterns).

Source: https://github.com/oconnor663/blake2_simd (MIT license)
