+++ crates/precompile/src/blake2/portable.rs
@@ -1,0 +1,52 @@
+// Adapted from https://github.com/oconnor663/blake2_simd (MIT license).
+// Portable BLAKE2b compression function, modified for EIP-152 variable rounds.
+
+use super::{IV, SIGMA};
+
+#[inline(always)]
+const fn g(v: &mut [u64; 16], a: usize, b: usize, c: usize, d: usize, x: u64, y: u64) {
+    v[a] = v[a].wrapping_add(v[b]).wrapping_add(x);
+    v[d] = (v[d] ^ v[a]).rotate_right(32);
+    v[c] = v[c].wrapping_add(v[d]);
+    v[b] = (v[b] ^ v[c]).rotate_right(24);
+    v[a] = v[a].wrapping_add(v[b]).wrapping_add(y);
+    v[d] = (v[d] ^ v[a]).rotate_right(16);
+    v[c] = v[c].wrapping_add(v[d]);
+    v[b] = (v[b] ^ v[c]).rotate_right(63);
+}
+
+#[inline(always)]
+const fn round(r: usize, m: &[u64; 16], v: &mut [u64; 16]) {
+    let s = SIGMA[r % 10];
+    // Mix the columns.
+    g(v, 0, 4, 8, 12, m[s[0]], m[s[1]]);
+    g(v, 1, 5, 9, 13, m[s[2]], m[s[3]]);
+    g(v, 2, 6, 10, 14, m[s[4]], m[s[5]]);
+    g(v, 3, 7, 11, 15, m[s[6]], m[s[7]]);
+    // Mix the rows.
+    g(v, 0, 5, 10, 15, m[s[8]], m[s[9]]);
+    g(v, 1, 6, 11, 12, m[s[10]], m[s[11]]);
+    g(v, 2, 7, 8, 13, m[s[12]], m[s[13]]);
+    g(v, 3, 4, 9, 14, m[s[14]], m[s[15]]);
+}
+
+pub(crate) fn compress(rounds: u32, h: &mut [u64; 8], m: &[u64; 16], t: &[u64; 2], f: bool) {
+    let mut v = [0u64; 16];
+    v[..8].copy_from_slice(h);
+    v[8..].copy_from_slice(&IV);
+
+    v[12] ^= t[0];
+    v[13] ^= t[1];
+
+    if f {
+        v[14] = !v[14];
+    }
+
+    for i in 0..rounds as usize {
+        round(i, m, &mut v);
+    }
+
+    for i in 0..8 {
+        h[i] ^= v[i] ^ v[i + 8];
+    }
+}
