+++ crates/precompile/src/interface.rs
@@ -255,7 +255,7 @@ pub trait Crypto: Send + Sync + Debug {
     /// Blake2 compression function.
     #[inline]
     fn blake2_compress(&self, rounds: u32, h: &mut [u64; 8], m: &[u64; 16], t: &[u64; 2], f: bool) {
-        crate::blake2::algo::compress(rounds as usize, h, m, t, f);
+        crate::blake2::compress(rounds, h, m, t, f);
     }
 
     /// secp256r1 (P-256) signature verification.
